# feat(sui-perf): add support to measure axios' requests