## Description
This is one part of a 3 PRs that solves [this issue](https://github.com/SUI-Components/sui/issues/79) :
* Add support to measure axios' requests ([sui](https://github.com/SUI-Components/sui/pull/111))
* Add request id to the domain on SSR ([mt--web-app](https://github.schibsted.io/scmspain/frontend-mt--web-app/pull/126))
* Add (commented) support for axios measurement ([mt--lib-motor](https://github.schibsted.io/scmspain/frontend-mt--lib-motor/pull/150))

## Additional notes
Testing this feature on mt--web-app with 3 simultaneous request and cache deactivated for news, we get the following:

```
timeline        start  time   %    label
███████████████ 0ms    4811ms 100% Request
 █              399ms  272ms  6%   CreateContext
  ████          672ms  1363ms 28%  UseCases
   ███          1101ms 934ms  19%  🌎  GET /https://ms-mt--api-web.spain.schibsted.io/advertisement-configuration #1
          ███   3236ms 979ms  20%  🌎  GET /https://ms-mt--api-web.spain.schibsted.io/news/home #2
          ███   3236ms 897ms  19%  🌎  GET /https://ms-mt--api-web.spain.schibsted.io/ads/home/ #5
          ████  3245ms 1345ms 28%  🌎  GET /https://ms-mt--api-web.spain.schibsted.io/tagcloud-groups #6
timeline        start  time   %    label
███████████████ 0ms    3203ms 100% Request
⸠               1ms    1ms    0%   CreateContext
████████        2ms    1613ms 50%  UseCases
        ████    1623ms 869ms  27%  🌎  GET /https://ms-mt--api-web.spain.schibsted.io/news/home #3
timeline        start  time   %    label
███████████████ 0ms    3317ms 100% Request
⸠               1ms    1ms    0%   CreateContext
███████         2ms    1611ms 49%  UseCases
       ███████  1619ms 1588ms 48%  🌎  GET /https://ms-mt--api-web.spain.schibsted.io/news/home #4
```

ISSUES CLOSED: #79

Please review @carlosvillu @miduga